+++ full_env_start.sh
@@ -211,8 +211,12 @@ fi
 # Only copy Xauthority if we're in a different directory than the old home
 if [ "$HOME" != "$OLD_HOME" ]; then
     echo "Copying Xauthority from $OLD_HOME to $HOME"
-    rm .Xauthority
+    if [ -f .Xauthority ]; then
+        rm .Xauthority
+        echo "Removed existing .Xauthority"
+    fi
     cp "$OLD_HOME"/.Xauthority .Xauthority
+    echo "Copied Xauthority from $OLD_HOME to $HOME"
 fi
 
 ############### Add useful alisas ###############
